[Unity] Give a better generic name to severals scripts and change indentation from default style on VS mac to VS "format code" function
